+++ src/main/java/com/saraiva/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import com.saraiva.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	public User findByUsername(String username);
 	public User findByUsernameAndPassword(String username, String password);
 	public User findByEmail(String email);
-	public User findById(long id);
+	public User findById(int id);
 }
