+++ src/main/java/com/saraiva/service/UserService.java
@@ -43,7 +43,7 @@ public class UserService {
 		}
 	}
 
-	public User findById(long id) {
+	public User findById(int id) {
 		return userRepository.findById(id);
 	}
 }
